Document Management and RAG-based Q&A Application/RAG/requirements.txt
fastapi
uvicorn
langchain
openai
pgvector
sqlalchemy
psycopg2-binary
python-dotenv